FarmqDomain/requirements.txt
pandas
streamlit
deep-translator
google-search-results
boto3
python-dotenv
sentence-transformers
torch
numpy
scikit-learn